RFP
--------------------------------------------------
Results of IcarusDev vs IcarusMain (8+0.08, NULL, 16MB, UHO_Lichess_4852_v1.epd):
Elo: 74.03 +/- 26.71, nElo: 94.67 +/- 33.07
LOS: 100.00 %, DrawRatio: 37.26 %, PairsRatio: 2.41
Games: 424, Wins: 204, Losses: 115, Draws: 105, Points: 256.5 (60.50 %)
Ptnml(0-2): [9, 30, 79, 51, 43], WL/DD Ratio: 5.58
LLR: 2.96 (100.5%) (-2.94, 2.94) [0.00, 10.00]
--------------------------------------------------

+++ src/search/searcher.rs
@@ -237,7 +237,7 @@ fn id_loop(mut pos: Position, thread: &mut ThreadCtx, print: bool) {
         thread.sel_depth = 0;
         let new_score = search::<Root>(
             &mut pos,
-            depth as i32,
+            depth as i16,
             0,
             -Score::INFINITE,
             Score::INFINITE,
